requirements.txt
# Core dependencies for ENSTRECT Multiview Segmentation
# Python 3.10+ required

# Deep Learning
torch>=2.5.0
torchvision>=0.20.0

# Image Processing
Pillow>=10.0.0
numpy>=1.26.0,<2.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# ENSTRECT dependencies (will be installed with enstrect)
# nnunetv2>=2.2
# pyntcloud>=0.3
# open3d>=0.17
# scikit-learn>=1.3
# networkx>=3.0
# alphashape>=1.3

# Optional: For 3D pipeline (requires PyTorch3D)
# pytorch3d>=0.7.0

# LLM + RAG Dependencies
transformers>=4.41.0
accelerate>=0.30.0
qwen-vl-utils
torch>=2.5.0

# RAG
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.5.0

# PDF Processing
pdfplumber>=0.10.0
pypdf2>=3.0.0

# Data Processing
jsonschema>=4.21.0
python-dateutil>=2.8.0
